requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-oidc==2.2.2
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.7.0
Werkzeug==3.1.3
